test(gmx): share mock-oracle and Lagoon valuation helpers
_set_mock_token_price() in the Lagoon PnL test module duplicated
setup_mock_oracle()'s raw ABI JSON parsing instead of using the shared
eth_defi.abi.get_contract() loader. Replaced with a reusable
set_mock_token_price() in eth_defi/gmx/testing/oracle.py.

test_valuation.py also imported fixtures/helpers directly from another
test module (tests.gmx.lagoon.test_gmx_close_pnl_token), which does not
work reliably under pytest's import mechanism. Moved the NAV
end-to-end test beside its Lagoon fixture and helpers instead.
